bugfix: 补齐采集模板只读反馈与回归
