+++ server/apps/monitor/tests/test_plugin_permission.py
@@ -8,7 +8,6 @@ from types import SimpleNamespace
 from unittest.mock import Mock, patch
 
 import pytest
-from rest_framework.test import APIRequestFactory, force_authenticate
 
 from apps.core.decorators.api_permission import HasPermission
 from apps.monitor.views.plugin import MonitorPluginViewSet
@@ -100,6 +99,18 @@ def test_只读权限_get_可读取采集模板():
     update_template.assert_not_called()
 
 
+def test_无关权限_get_在读取插件前被拒绝():
+    result, get_object, get_template, update_template = _call_collect_template(
+        "GET",
+        {"some_other-View"},
+    )
+
+    assert result == _DENIED
+    get_object.assert_not_called()
+    get_template.assert_not_called()
+    update_template.assert_not_called()
+
+
 def test_写权限_put_可更新采集模板():
     result, get_object, get_template, update_template = _call_collect_template(
         "PUT",
@@ -136,75 +147,3 @@ def test_超级用户_put_继续兼容更新采集模板():
     get_object.assert_called_once_with()
     get_template.assert_not_called()
     update_template.assert_called_once()
-
-
-def test_只读权限_http_put_在读取插件前返回_403():
-    request = APIRequestFactory().put(
-        "/monitor/api/monitor_plugin/plugin-1/collect_template/",
-        {"content": "[[inputs.snmp.field]]"},
-        format="json",
-    )
-    force_authenticate(
-        request,
-        user=SimpleNamespace(
-            is_authenticated=True,
-            is_superuser=False,
-            permission={"monitor": {"integration_collect-View"}},
-            locale="en",
-        ),
-    )
-
-    with (
-        patch.object(MonitorPluginViewSet, "get_object") as get_object,
-        patch(
-            "apps.monitor.views.plugin.CustomSnmpPluginService.update_collect_template",
-            return_value={"content": "new"},
-        ) as update_template,
-    ):
-        response = MonitorPluginViewSet.as_view({"put": "collect_template"})(request, pk="plugin-1")
-
-    assert response.status_code == 403
-    get_object.assert_not_called()
-    update_template.assert_not_called()
-
-
-def test_写权限_http_head_沿用读取路径且不更新模板():
-    request = APIRequestFactory().head(
-        "/monitor/api/monitor_plugin/plugin-1/collect_template/",
-    )
-    force_authenticate(
-        request,
-        user=SimpleNamespace(
-            is_authenticated=True,
-            is_superuser=False,
-            permission={"monitor": {"integration_configure-Add"}},
-            locale="en",
-        ),
-    )
-
-    with (
-        patch.object(
-            MonitorPluginViewSet,
-            "get_object",
-            return_value=SimpleNamespace(template_type="snmp"),
-        ) as get_object,
-        patch(
-            "apps.monitor.views.plugin.CustomSnmpPluginService.get_collect_template",
-            return_value={"content": "old"},
-        ) as get_template,
-        patch(
-            "apps.monitor.views.plugin.CustomSnmpPluginService.update_collect_template",
-            return_value={"content": "new"},
-        ) as update_template,
-    ):
-        response = MonitorPluginViewSet.as_view(
-            {
-                "get": "collect_template",
-                "put": "collect_template",
-            }
-        )(request, pk="plugin-1")
-
-    assert response.status_code == 200
-    get_object.assert_called_once_with()
-    get_template.assert_called_once()
-    update_template.assert_not_called()
